feat (front-end SDK and DO):
Implemented a patch object on the DO backend that is passed to the front end and used to implement delta updates.

Added logic on the frontend to add patch updates to the localState variable from the Syncosaurus class.

Removed canonState instance variable from the Syncosaurus class since it was no longer required after switching to delta updates.

Co-authored-by: Alex Sklar <sklarfox@gmail.com>
Co-authored-by: Erik Margetis <erik.margetis@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>