feat(presence): Basic data flow for presence, hard coded data
Set up the data flow for presence. Presence is considered separate
from canon state, and is on its own property that is to be updated
live as users connect/disconnect (ephemeral data).

Added in scaffolding for the usePresence hook.

The new property has been added on both the Syncosaurus client class
as well as in the durable object. Basic checks/subscritions are set
up such that the app re-renders appropriately.

Co-authored-by: Erik Margetis <erik.margetis@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>
Co-authored-by: Franco Waisfeld <Francowaisfeld@gmail.com>