feat (front-end SDK):
Updated the subscribe methods to track multiple keys that are used in a "query" passed to the useSubscribe method

Added a queryTransaction class which has methods `get`, `has` and `isEmpty` which can be used inside of a query. The queryTransaction class also keeps track of methods executed to create a list of multiple keys that a query depends on.

Updated App.jsx to use a query instead of a single property name.

Co-authored-by: Alex Sklar <sklarfox@gmail.com>
Co-authored-by: Franco Waisfeld <francowaisfeld@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>