feat(front-end SDK and counter demo app): Create a componenets folder for the demo app and seperated the counter into its own component.
Added counter buttons that display room names. Clicking on a button makes the associated counter appear on screen.

Modified the Syncosaurus class in the front-end sdk to take a roomID and pass it to the cloudflare worker so that it can establish a websocket connection between the client and the asssociated durable object.

Seperated all logic that existed in the constructor function for intialzing a websocket connection and mutator functions into their own methods.

Co-authored-by: Alex Sklar <sklarfox@gmail.com>
Co-authored-by: Erik Margetis <erik.margetis@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>