feat (Frontend SDK and Backend / DO):
Implemented sending the snapshot ID on init which is one less than current snapshot ID. This will allow the client to catch the first delta if it goes missing.

Implemented sending a reset type which will send the whole cannon, similar to . This logic could be upgraded to send missing deltas if we decided to preserve them in the future.

Implemented different `updateType`s on server side to help client differentiate between message types rather than relying on present or missing properties as a way to determine what to do with the message.

Implemented client-side logic to handle all three `updateType`s: `reset`, `init`, `delta`

Co-authored-by: Alex Sklar <sklarfox@gmail.com>
Co-authored-by: Franco Waisfeld <francowaisfeld@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>