feat(Frontend SDK and DO):
The scan method was implemented on the `ReadTransaction` class on the frontend and the `ServerTransaction` class on the backend. The scan method takes a developer-defined callback function and returns an object of key-value pairs where that callback evaluates to `true`. `scan` can be thought of a way to retrieve a filtered version of the local KV store based on criteria for each key or its associated value.

To test the `scan` method, a simple to-do list was implemented. The app allows users to add a text to a list and and delete it. The state will be updated across clients accordingly.

Co-authored-by: Alex Sklar <sklarfox@gmail.com>
Co-authored-by: Franco Waisfeld <francowaisfeld@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>