feat(autosave): add autosave parameter, and  conditional check if storage is used during construction
If the USE_STORAGE env variable is set to true, then the constructor will create a setInterval (also set with the env variable) what saves the current `canon` state every X seconds.

Additionally, if USE_STORAGE is true, then the constructor will load the previously saved canon state and reassign the state on the websocket server.

Co-authored-by: Erik Margetis <erik.margetis@gmail.com>
Co-authored-by: Joseph Liang <j3liang@icloud.com>
Co-authored-by: Franco Waisfeld <Francowaisfeld@gmail.com>